Fix comments after review: moved logic methods from helper_order.py to contest.py. Rename and moved helpers.py. README file updated accordingly.
