+++ tests/test_courier_login.py
@@ -2,7 +2,7 @@ import pytest
 import allure
 
 from data import ResponseMessage as RM, ResponseStatus as RS, CourierData
-from helpers.helper_courier import HelperCourier
+from helpers import HelperCourier
 from api.courier_api import CourierAPI
