+++ tests/test_order_accept.py
@@ -1,10 +1,8 @@
 import allure
-import pytest
 
 from data import ResponseMessage as RM, ResponseStatus as RS, OrderData, CourierData
 from api.order_api import OrderApi
-from helpers.helper_courier import HelperCourier
-from helpers.helper_order import HelperOrder
+from helpers import HelperCourier
 
 
 class TestOrderAccept:
@@ -13,9 +11,9 @@ class TestOrderAccept:
         'Verify that the order is accepted successfully if passing valid order and courier id.')
     @allure.description(
         'Verify that 200 code is returned for PUT request with valid order and courier id.')
-    def test_accept_order_accepted_successfully(self, create_courier_and_return_courier_id):
-        order_id = HelperOrder.get_order_id()
-        response = OrderApi.accept_order(order_id, create_courier_and_return_courier_id)
+    def test_accept_order_accepted_successfully(self, create_courier_and_return_courier_id,
+                                                create_order_and_return_order_id):
+        response = OrderApi.accept_order(create_order_and_return_order_id, create_courier_and_return_courier_id)
         assert response.status_code == RS.OK and response.text == RM.SUCCESSFULL_ACCEPT_ORDER
         HelperCourier.delete_courier(create_courier_and_return_courier_id)
 
@@ -23,10 +21,9 @@ class TestOrderAccept:
         'Verify that the order is not accepted if courier id value is not passed.')
     @allure.description(
         'Verify that 400 code is returned for PUT request with empty value for courier id.')
-    def test_accept_order_no_courier_id_not_accepted(self):
+    def test_accept_order_no_courier_id_not_accepted(self, create_order_and_return_order_id):
         courier_id = ''
-        order_id = HelperOrder.get_order_id()
-        response = OrderApi.accept_order(order_id, courier_id)
+        response = OrderApi.accept_order(create_order_and_return_order_id, courier_id)
         assert response.status_code == RS.BAD_REQUEST and response.json()['message'] == RM.ACCEPT_ORDER_MISSED_PARAMETER
 
     @allure.title(
@@ -51,9 +48,8 @@ class TestOrderAccept:
         'Verify that the order is not accepted if passing invalid courier id.')
     @allure.description(
         'Verify that 404 code is returned for PUT request if courier id is invalid.')
-    def test_accept_order_not_existing_courier_id_not_accepted(self):
-        order_id = HelperOrder.get_order_id()
-        response = OrderApi.accept_order(order_id, CourierData.NOT_EXISTING_COURIER_ID)
+    def test_accept_order_not_existing_courier_id_not_accepted(self, create_order_and_return_order_id):
+        response = OrderApi.accept_order(create_order_and_return_order_id, CourierData.NOT_EXISTING_COURIER_ID)
         assert response.status_code == RS.NOT_FOUND and response.json()['message'] == RM.ACCEPT_ORDER_INVALID_COURIER_ID
 
     @allure.title(
